// src/types/data.ts
export interface IData {
  emojis: string[];
  greeting: IGreeting;
  hostInfo: IHostInfo[];
  locationInfo: ILocationInfo[];
  mapInfo: IMapInfo;
}

export interface IGreeting {
  mainImg: string;
  title: string;
  date: string;
  message: string;
  host: {
    groom: {
      name: string;
      relation: string;
      parents: Parent;
    };
    bride: {
      name: string;
      relation: string;
      parents: Parent;
    };
  };
  eventDetail: string;
}

export interface BrideAndGroom {
  name: string;
  relation: string;
  parents: Parent;
  origin: string;
}

type Parent = { relation: string; isDeceased?: boolean; name: string }[];

export interface IHostInfo {
  host: string;
  accountInfo: { name: string; relation: string; bank: string; account: string }[];
}

export interface IMapInfo {
  address1: string;
  address2: string;
  naverMap: string;
  kakaoMap: string;
  lat: number;
  lon: number;
}

export interface ILocationInfo {
  title: string;
  desc: string;
}